Added extra argument to match latest script version.
When building with xcodebuild command, running this script will cause the build to fail since it produces a verbose output. Both Xcode IDE and xcodebuild consider that there are errors in that output, but where Xcode ignores them, xcodebuild tool fails to build.

This extra argument allows us to call the script with a flag to include a less verbose output, allowing us to succesfully run the xcodebuild command tool successfully.